manage all  order delete fix

+++ src/Pages/Dashboard/ManageAllOrders/ManageAllOrders.js
@@ -20,7 +20,7 @@ const ManageAllOrders = () => {
     // console.log(status);
 
     const handleUpdateStatus = (id) => {
-        fetch(`https://nameless-depths-17324.herokuapp.com/updateStatus/${id}`, {
+        fetch(`http://localhost:5000/updateStatus/${id}`, {
             method: "PUT",
             headers: { "content-type": "application/json" },
             body: JSON.stringify({ status }),
@@ -29,19 +29,22 @@ const ManageAllOrders = () => {
         console.log(id);
     };
 
-    const handleDelete = (id, email) => {
-        fetch(`https://nameless-depths-17324.herokuapp.com/deleteOrder?id=${id}&email=${email}`, {
-            method: "DELETE",
-        })
-            .then(res => res.json())
-            .then(data => {
-                // if (data.deleteCount) {
-                //     setControl(!control);
-                // }
-                console.log(data);
-            });
-
-        // console.log(id, email);
+    const handleDelete = (id) => {
+        const proceed = window.confirm("Do you really want to delete?");
+        if (proceed) {
+            fetch(`http://localhost:5000/deleteOrder/${id}`, {
+                method: "DELETE",
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
+                        setControl(true);
+                        const remainingOrders = orders.filter(order => order._id !== id);
+                        setOrders(remainingOrders);
+                    }
+                });
+        }
+        // console.log(id);
     };
 
     return (
@@ -54,10 +57,11 @@ const ManageAllOrders = () => {
                         <div >
                             <h4 >{one.name}</h4>
                             <h5>Email: {one.email}</h5>
+                            <h6>Date: {one.date}</h6>
                             <h5>${one.price}</h5>
                             Status : <input onChange={handleStatus} defaultValue={one.status} />(update to 'Shipped')
                             <button onClick={() => handleUpdateStatus(one._id)} >Update</button><br />
-                            <button onClick={() => handleDelete(one._id, one.email)} >Delete</button>
+                            <button onClick={() => handleDelete(one._id)} >Delete</button>
                         </div>
                     </div>
                 )}
